# Performance changes and updates for Julia v1.11