--- a/.gitignore
+++ b/.gitignore
@@ -24,3 +24,5 @@ docs/site/
 Manifest.toml
 
 .vscode/
+
+sandbox