--- a/src/DamNode.jl
+++ b/src/DamNode.jl
@@ -161,25 +161,25 @@ Update dam volume for timestep.
 volume of water stored in dam
 """
 function update_volume(volume, node_inflow, gamma, rain, evap, area, extractions, discharge, max_store)::Float64
-    
+
     vol = volume + (node_inflow + gamma) + (rain - evap) * area - extractions - discharge
 
     return max(0.0, min(max_store, vol))
 end
 
 
-function run_node!(node::DamNode, climate::Climate; 
+function run_node!(node::DamNode, climate::Climate;
                    inflow=nothing, extraction=nothing, exchange=nothing)
     timesteps = sim_length(climate)
     for ts in 1:timesteps
-        run_node!(node, climate, ts; 
+        run_node!(node, climate, ts;
                   inflow=inflow, extraction=extraction, exchange=exchange)
     end
 end
 
 
 """
-    run_node!(node::DamNode, climate::Climate, timestep::Int; 
+    run_node!(node::DamNode, climate::Climate, timestep::Int;
               inflow=nothing, extraction=nothing, exchange=nothing)
 
 Run a specific node for a specified time step.
@@ -192,7 +192,7 @@ Run a specific node for a specified time step.
 - `extraction::DataFrame` : Time series of water orders (expects column of `_releases`)
 - `exchange::DataFrame` : Time series of groundwater flux
 """
-function run_node!(node::DamNode, climate::Climate, timestep::Int; 
+function run_node!(node::DamNode, climate::Climate, timestep::Int;
                    inflow=nothing, extraction=nothing, exchange=nothing)
     ts = timestep
     if checkbounds(Bool, node.outflow, ts)
@@ -220,14 +220,14 @@ Calculate outflow for the dam node for a single time step.
 - node : DamNode
 - rain : rainfall in mm
 - et : evapotranspiration data in mm
-- irrig_ext : irrigation extractions 
+- irrig_ext : irrigation extractions
 - extractions : extraction data in ML
 - gw_flux : groundwater interaction
 
 # Returns
 - outflow from dam
 """
-function run_node!(node::DamNode, 
+function run_node!(node::DamNode,
                    rain::Float64,
                    et::Float64,
                    vol::Float64,