--- a/src/GR4JNode.jl
+++ b/src/GR4JNode.jl
@@ -55,12 +55,12 @@ A four-parameter model with two stores.
 - x4 : time base of unit hydrograph UH1 (days, > 0.5)
 
 # References
-1. Perrin, C., Michel, C., Andréassian, V., 2003. 
-    Improvement of a parsimonious model for streamflow simulation. 
+1. Perrin, C., Michel, C., Andréassian, V., 2003.
+    Improvement of a parsimonious model for streamflow simulation.
     Journal of Hydrology 279, 275-289.
     https://doi.org/10.1016/S0022-1694(03)00225-7
 
-2. MacDonald, A. 2014. 
+2. MacDonald, A. 2014.
     Python GR4J
     https://github.com/amacd31/gr4j