```julia
# HyMod
julia> @time run_node!(hymod_node, climate);
  0.075078 seconds (1.86 M allocations: 45.034 MiB, 12.45% gc time)

# After optimizations
julia> @time run_node!(hymod_node, climate);
  0.017009 seconds (395.05 k allocations: 12.057 MiB)

  
# SYMHYD
julia> @time run_node!(symhyd_node, climate)
  0.094059 seconds (2.09 M allocations: 50.622 MiB, 12.94% gc time)
  
# After optimizations
julia> @time run_node!(symhyd_node, climate);
  0.020182 seconds (656.82 k allocations: 15.046 MiB)
  

# GR4J
julia> @time run_node!(gr4j_node, climate);
  0.061611 seconds (1.18 M allocations: 39.326 MiB)

# After Optimizations

julia> @time run_node!(gr4j_node, climate);
  0.015513 seconds (150.55 k allocations: 4.451 MiB)
  

# IHACRES
julia> @time run_node!(ihacres_node, climate);
  0.159082 seconds (1.50 M allocations: 47.337 MiB, 7.26% gc time)
  
# After Optimizations

julia> @time run_node!(ihacres_node, climate);
  0.108612 seconds (902.87 k allocations: 23.967 MiB)
```

Note the `ModelParameters.jl` dependency currently has a bug. A fix is in the works but for the moment, use the `fix-62-array-methoderror` branch.